+++ src/lib.rs
@@ -4,26 +4,22 @@ pub use input::*;
 
 use std::io::{self, BufReader, Read, Write};
 
-use anstyle::{AnsiColor, Reset};
+use owo_colors::{colors, Color};
 
-#[derive(PartialEq, Eq)]
-#[repr(usize)]
-enum ByteColor {
-    Null = 0,
-    Offset = 1,
-    AsciiPrintable = 2,
-    AsciiWhitespace = 3,
-    AsciiOther = 4,
-    NonAscii = 5,
-    Reset = 6,
+pub enum Base {
+    Binary,
+    Octal,
+    Decimal,
+    Hexadecimal,
 }
 
-const COLOR_NULL: AnsiColor = AnsiColor::BrightBlack;
-const COLOR_OFFSET: AnsiColor = AnsiColor::BrightBlack;
-const COLOR_ASCII_PRINTABLE: AnsiColor = AnsiColor::Cyan;
-const COLOR_ASCII_WHITESPACE: AnsiColor = AnsiColor::Green;
-const COLOR_ASCII_OTHER: AnsiColor = AnsiColor::Magenta;
-const COLOR_NONASCII: AnsiColor = AnsiColor::Yellow;
+const COLOR_NULL: &[u8] = colors::BrightBlack::ANSI_FG.as_bytes();
+const COLOR_OFFSET: &[u8] = colors::BrightBlack::ANSI_FG.as_bytes();
+const COLOR_ASCII_PRINTABLE: &[u8] = colors::Cyan::ANSI_FG.as_bytes();
+const COLOR_ASCII_WHITESPACE: &[u8] = colors::Green::ANSI_FG.as_bytes();
+const COLOR_ASCII_OTHER: &[u8] = colors::Green::ANSI_FG.as_bytes();
+const COLOR_NONASCII: &[u8] = colors::Yellow::ANSI_FG.as_bytes();
+const COLOR_RESET: &[u8] = colors::Default::ANSI_FG.as_bytes();
 
 pub enum ByteCategory {
     Null,
@@ -59,15 +55,15 @@ impl Byte {
         }
     }
 
-    fn color(self) -> ByteColor {
+    fn color(self) -> &'static [u8] {
         use crate::ByteCategory::*;
 
         match self.category() {
-            Null => ByteColor::Null,
-            AsciiPrintable => ByteColor::AsciiPrintable,
-            AsciiWhitespace => ByteColor::AsciiWhitespace,
-            AsciiOther => ByteColor::AsciiOther,
-            NonAscii => ByteColor::NonAscii,
+            Null => COLOR_NULL,
+            AsciiPrintable => COLOR_ASCII_PRINTABLE,
+            AsciiWhitespace => COLOR_ASCII_WHITESPACE,
+            AsciiOther => COLOR_ASCII_OTHER,
+            NonAscii => COLOR_NONASCII,
         }
     }
 
@@ -75,7 +71,7 @@ impl Byte {
         use crate::ByteCategory::*;
 
         match self.category() {
-            Null => '0',
+            Null => '⋄',
             AsciiPrintable => self.0 as char,
             AsciiWhitespace if self.0 == 0x20 => ' ',
             AsciiWhitespace => '_',
@@ -161,7 +157,8 @@ pub struct PrinterBuilder<'a, Writer: Write> {
     border_style: BorderStyle,
     use_squeeze: bool,
     panels: u64,
-    group_bytes: u8,
+    group_size: u8,
+    base: Base,
 }
 
 impl<'a, Writer: Write> PrinterBuilder<'a, Writer> {
@@ -174,7 +171,8 @@ impl<'a, Writer: Write> PrinterBuilder<'a, Writer> {
             border_style: BorderStyle::Unicode,
             use_squeeze: true,
             panels: 2,
-            group_bytes: 1,
+            group_size: 1,
+            base: Base::Hexadecimal,
         }
     }
 
@@ -208,8 +206,13 @@ impl<'a, Writer: Write> PrinterBuilder<'a, Writer> {
         self
     }
 
-    pub fn num_group_bytes(mut self, num: u8) -> Self {
-        self.group_bytes = num;
+    pub fn group_size(mut self, num: u8) -> Self {
+        self.group_size = num;
+        self
+    }
+
+    pub fn with_base(mut self, base: Base) -> Self {
+        self.base = base;
         self
     }
 
@@ -222,7 +225,8 @@ impl<'a, Writer: Write> PrinterBuilder<'a, Writer> {
             self.border_style,
             self.use_squeeze,
             self.panels,
-            self.group_bytes,
+            self.group_size,
+            self.base,
         )
     }
 }
@@ -235,21 +239,21 @@ pub struct Printer<'a, Writer: Write> {
     show_char_panel: bool,
     show_position_panel: bool,
     show_color: bool,
-    colors: Vec<String>,
-    curr_color: Option<ByteColor>,
+    curr_color: Option<&'static [u8]>,
     border_style: BorderStyle,
     byte_hex_panel: Vec<String>,
     byte_char_panel: Vec<String>,
     // same as previous but in Fixed(242) gray color, for position panel
     byte_hex_panel_g: Vec<String>,
-    byte_char_panel_g: Vec<String>,
     squeezer: Squeezer,
     display_offset: u64,
     /// The number of panels to draw.
     panels: u64,
     squeeze_byte: usize,
     /// The number of octets per group.
-    group_bytes: u8,
+    group_size: u8,
+    /// The number of digits used to write the base.
+    base_digits: u8,
 }
 
 impl<'a, Writer: Write> Printer<'a, Writer> {
@@ -261,7 +265,8 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
         border_style: BorderStyle,
         use_squeeze: bool,
         panels: u64,
-        group_bytes: u8,
+        group_size: u8,
+        base: Base,
     ) -> Printer<'a, Writer> {
         Printer {
             idx: 0,
@@ -271,24 +276,19 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
             show_position_panel,
             show_color,
             curr_color: None,
-            colors: vec![
-                COLOR_NULL.render_fg().to_string(),
-                COLOR_OFFSET.render_fg().to_string(),
-                COLOR_ASCII_PRINTABLE.render_fg().to_string(),
-                COLOR_ASCII_WHITESPACE.render_fg().to_string(),
-                COLOR_ASCII_OTHER.render_fg().to_string(),
-                COLOR_NONASCII.render_fg().to_string(),
-                Reset.render().to_string(),
-            ],
             border_style,
-            byte_hex_panel: (0u8..=u8::MAX).map(|i| format!("{:02x}", i)).collect(),
+            byte_hex_panel: (0u8..=u8::MAX)
+                .map(|i| match base {
+                    Base::Binary => format!("{i:08b}"),
+                    Base::Octal => format!("{i:03o}"),
+                    Base::Decimal => format!("{i:03}"),
+                    Base::Hexadecimal => format!("{i:02x}"),
+                })
+                .collect(),
             byte_char_panel: (0u8..=u8::MAX)
                 .map(|i| format!("{}", Byte(i).as_char()))
                 .collect(),
-            byte_hex_panel_g: (0u8..=u8::MAX).map(|i| format!("{:02x}", i)).collect(),
-            byte_char_panel_g: (0u8..=u8::MAX)
-                .map(|i| format!("{}", Byte(i).as_char()))
-                .collect(),
+            byte_hex_panel_g: (0u8..=u8::MAX).map(|i| format!("{i:02x}")).collect(),
             squeezer: if use_squeeze {
                 Squeezer::Ignore
             } else {
@@ -297,7 +297,13 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
             display_offset: 0,
             panels,
             squeeze_byte: 0x00,
-            group_bytes,
+            group_size,
+            base_digits: match base {
+                Base::Binary => 8,
+                Base::Octal => 3,
+                Base::Decimal => 3,
+                Base::Hexadecimal => 2,
+            },
         }
     }
 
@@ -308,8 +314,8 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
 
     fn panel_sz(&self) -> usize {
         // add one to include the trailing space of a group
-        let group_sz = 2 * self.group_bytes as usize + 1;
-        let group_per_panel = 8 / self.group_bytes as usize;
+        let group_sz = self.base_digits as usize * self.group_size as usize + 1;
+        let group_per_panel = 8 / self.group_size as usize;
         // add one to include the leading space
         1 + group_sz * group_per_panel
     }
@@ -323,27 +329,27 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
         let h_repeat = h.to_string().repeat(self.panel_sz());
 
         if self.show_position_panel {
-            write!(self.writer, "{l}{h8}{c}", l = l, c = c, h8 = h8).ok();
+            write!(self.writer, "{l}{h8}{c}")?;
         } else {
-            write!(self.writer, "{}", l).ok();
+            write!(self.writer, "{l}")?;
         }
 
         for _ in 0..self.panels - 1 {
-            write!(self.writer, "{h_repeat}{c}", h_repeat = h_repeat, c = c).ok();
+            write!(self.writer, "{h_repeat}{c}")?;
         }
         if self.show_char_panel {
-            write!(self.writer, "{h_repeat}{c}", h_repeat = h_repeat, c = c).ok();
+            write!(self.writer, "{h_repeat}{c}")?;
         } else {
-            write!(self.writer, "{h_repeat}", h_repeat = h_repeat).ok();
+            write!(self.writer, "{h_repeat}")?;
         }
 
         if self.show_char_panel {
             for _ in 0..self.panels - 1 {
-                write!(self.writer, "{h8}{c}", h8 = h8, c = c).ok();
+                write!(self.writer, "{h8}{c}")?;
             }
-            writeln!(self.writer, "{h8}{r}", h8 = h8, r = r).ok();
+            writeln!(self.writer, "{h8}{r}")?;
         } else {
-            writeln!(self.writer, "{r}", r = r).ok();
+            writeln!(self.writer, "{r}")?;
         }
 
         Ok(())
@@ -371,17 +377,15 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
                 .as_bytes(),
         )?;
         if self.show_color {
-            self.writer
-                .write_all(self.colors[ByteColor::Offset as usize].as_bytes())?;
+            self.writer.write_all(COLOR_OFFSET)?;
         }
         if self.show_position_panel {
             match self.squeezer {
                 Squeezer::Print => {
                     self.writer
-                        .write_all(self.byte_char_panel_g[b'*' as usize].as_bytes())?;
+                        .write_all(self.byte_char_panel[b'*' as usize].as_bytes())?;
                     if self.show_color {
-                        self.writer
-                            .write_all(self.colors[ByteColor::Reset as usize].as_bytes())?;
+                        self.writer.write_all(COLOR_RESET)?;
                     }
                     self.writer.write_all(b"       ")?;
                 }
@@ -396,8 +400,7 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
                             .write_all(self.byte_hex_panel_g[byte as usize].as_bytes())?;
                     }
                     if self.show_color {
-                        self.writer
-                            .write_all(self.colors[ByteColor::Reset as usize].as_bytes())?;
+                        self.writer.write_all(COLOR_RESET)?;
                     }
                 }
             }
@@ -417,8 +420,7 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
             Squeezer::Ignore | Squeezer::Disabled => {
                 if let Some(&b) = self.line_buf.get(i as usize) {
                     if self.show_color && self.curr_color != Some(Byte(b).color()) {
-                        self.writer
-                            .write_all(self.colors[Byte(b).color() as usize].as_bytes())?;
+                        self.writer.write_all(Byte(b).color())?;
                         self.curr_color = Some(Byte(b).color());
                     }
                     self.writer
@@ -430,8 +432,7 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
         }
         if i == 8 * self.panels - 1 {
             if self.show_color {
-                self.writer
-                    .write_all(self.colors[ByteColor::Reset as usize].as_bytes())?;
+                self.writer.write_all(COLOR_RESET)?;
                 self.curr_color = None;
             }
             self.writer.write_all(
@@ -442,8 +443,7 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
             )?;
         } else if i % 8 == 7 {
             if self.show_color {
-                self.writer
-                    .write_all(self.colors[ByteColor::Reset as usize].as_bytes())?;
+                self.writer.write_all(COLOR_RESET)?;
                 self.curr_color = None;
             }
             self.writer.write_all(
@@ -469,30 +469,27 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
             Squeezer::Print => {
                 if !self.show_position_panel && i == 0 {
                     if self.show_color {
-                        self.writer
-                            .write_all(self.colors[ByteColor::Offset as usize].as_bytes())?;
+                        self.writer.write_all(COLOR_OFFSET)?;
                     }
                     self.writer
-                        .write_all(self.byte_char_panel_g[b'*' as usize].as_bytes())?;
+                        .write_all(self.byte_char_panel[b'*' as usize].as_bytes())?;
                     if self.show_color {
-                        self.writer
-                            .write_all(self.colors[ByteColor::Reset as usize].as_bytes())?;
+                        self.writer.write_all(COLOR_RESET)?;
                     }
-                } else {
-                    if i % (self.group_bytes as usize) == 0 {
-                        self.writer.write_all(b" ")?;
-                    }
+                } else if i % (self.group_size as usize) == 0 {
+                    self.writer.write_all(b" ")?;
                 }
-                self.writer.write_all(b"  ")?;
+                for _ in 0..self.base_digits {
+                    self.writer.write_all(b" ")?;
+                }
             }
             Squeezer::Delete => self.writer.write_all(b"   ")?,
             Squeezer::Ignore | Squeezer::Disabled => {
-                if i % (self.group_bytes as usize) == 0 {
+                if i % (self.group_size as usize) == 0 {
                     self.writer.write_all(b" ")?;
                 }
                 if self.show_color && self.curr_color != Some(Byte(b).color()) {
-                    self.writer
-                        .write_all(self.colors[Byte(b).color() as usize].as_bytes())?;
+                    self.writer.write_all(Byte(b).color())?;
                     self.curr_color = Some(Byte(b).color());
                 }
                 self.writer
@@ -503,8 +500,7 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
         if i % 8 == 7 {
             if self.show_color {
                 self.curr_color = None;
-                self.writer
-                    .write_all(self.colors[ByteColor::Reset as usize].as_bytes())?;
+                self.writer.write_all(COLOR_RESET)?;
             }
             self.writer.write_all(b" ")?;
             // byte is last in last panel
@@ -541,13 +537,15 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
 
         let mut buf = BufReader::new(reader);
 
-        self.print_header()?;
         let leftover = loop {
             // read a maximum of 8 * self.panels bytes from the reader
             if let Ok(n) = buf.read(&mut self.line_buf) {
                 if n > 0 && n < 8 * self.panels as usize {
                     // if less are read, that indicates end of file after
-                    is_empty = false;
+                    if is_empty {
+                        self.print_header()?;
+                        is_empty = false;
+                    }
 
                     // perform second check on read
                     if let Ok(0) = buf.read(&mut self.line_buf[n..]) {
@@ -564,7 +562,10 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
                     break None;
                 }
             }
-            is_empty = false;
+            if is_empty {
+                self.print_header()?;
+                is_empty = false;
+            }
 
             // squeeze is active, check if the line is the same
             // skip print if still squeezed, otherwise print and deactivate squeeze
@@ -617,6 +618,8 @@ impl<'a, Writer: Write> Printer<'a, Writer> {
         // special ending
 
         if is_empty {
+            self.base_digits = 2;
+            self.print_header()?;
             if self.show_position_panel {
                 write!(self.writer, "{0:9}", "│")?;
             }
@@ -678,6 +681,7 @@ mod tests {
             true,
             2,
             1,
+            Base::Hexadecimal,
         );
 
         printer.print_all(input).unwrap();
@@ -731,6 +735,7 @@ mod tests {
             true,
             2,
             1,
+            Base::Hexadecimal,
         );
         printer.display_offset(0xdeadbeef);
 
@@ -763,6 +768,7 @@ mod tests {
             true,
             4,
             1,
+            Base::Hexadecimal,
         );
 
         printer.print_all(input).unwrap();
@@ -776,9 +782,9 @@ mod tests {
         let input = io::Cursor::new(b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00");
         let expected_string = "\
 ┌────────┬─────────────────────────┬─────────────────────────┬────────┬────────┐
-│00000000│ 00 00 00 00 00 00 00 00 ┊ 00 00 00 00 00 00 00 00 │00000000┊00000000│
+│00000000│ 00 00 00 00 00 00 00 00 ┊ 00 00 00 00 00 00 00 00 │⋄⋄⋄⋄⋄⋄⋄⋄┊⋄⋄⋄⋄⋄⋄⋄⋄│
 │*       │                         ┊                         │        ┊        │
-│00000020│ 00                      ┊                         │0       ┊        │
+│00000020│ 00                      ┊                         │⋄       ┊        │
 └────────┴─────────────────────────┴─────────────────────────┴────────┴────────┘
 "
         .to_owned();
@@ -821,6 +827,7 @@ mod tests {
             true,
             3,
             1,
+            Base::Hexadecimal,
         );
 
         printer.print_all(input).unwrap();