+++ CHANGELOG.md
@@ -1,24 +1,47 @@
-# unreleased
-
-## Changes
-
+# v0.12.0
 
 ## Features
 
-- Breaking change (library): `Printer::new` is deprecated as a part of the public API. Alternatively, you can now construct a `Printer` using the `PrinterBuilder` builder API, see [#168](https://github.com/sharkdp/hexyl/pull/168). (@sharifhsn)
-
-- Added variable panels through the `--panels` and `--terminal-width` flags, see [#13](https://github.com/sharkdp/hexyl/issues/13) and [#164](https://github.com/sharkdp/hexyl/pull/164) (@sharifhsn)
-
-- Added new `--group-bytes`/`-g` option, see #104 and #170 (@RinHizakura)
+- Only show one panel by default if the terminal width is not wide enough for two panels, see #182 (@sharkdp)
+- Respect the `NO_COLOR` environment variable, see #179 (@sharifhsn)
 
 ## Bugfixes
 
+- Do not fail with an error if `--panels=auto` is used and the output is piped, see #184 (@sharkdp)
+
+## Changes
+
+- Breaking: For `xxd`-compatibility reasons, `--group_bytes` has been renamed to `--group-size` (with an `--groupsize` alias), see #121 (@sharkdp)
+
+## `hexyl` as a library
+
+- Breaking: `num_group_bytes` has been renamed to `group_size`.
+
+
+# v0.11.0
+
+## Features
+
+- Significantly improved performance, see #173 and #176 (@sharifhsn)
+- Added variable panels through the `--panels` and `--terminal-width` flags, see [#13](https://github.com/sharkdp/hexyl/issues/13) and [#164](https://github.com/sharkdp/hexyl/pull/164) (@sharifhsn)
+- Added new `--group-bytes`/`-g` option, see #104 and #170 (@RinHizakura)
+- Added new `--base B` option (where `B` can be `binary`, `octal`, `decimal` or `hexadecimal`), see #147 and #178 (@sharifhsn)
+- Show actual zero bytes as `⋄` in the character panel (previously: `0`), in order not to confuse them with ASCII
+  `0` bytes if colors are deactivated. Closes #166 (@sharkdp)
+
+## `hexyl` as a library
+
+- Breaking change: `Printer::new` is deprecated as a part of the public API. Alternatively, you can now construct a `Printer` using the `PrinterBuilder` builder API, see [#168](https://github.com/sharkdp/hexyl/pull/168). (@sharifhsn)
 
 ## Other
 
+- More tests for the squeezing feature, see #177 (@mkatychev)
 
-## Packaging
+## Thank you
 
+Special thanks go to @sharifhsn, not just for the new features,
+bugfixes and performance improvements. But also for many internal
+improvements of the code base and other maintenance tasks.
 
 
 # v0.10.0